requirements.txt
requests
beautifulsoup4
Pillow
opencv-python
pyzbar
numpy
